pyproject.toml
[tool.poetry]
name = "pepsflow"
version = "0.1.0"
description = ""
authors = ["DerkNiessink <derk@niessink.com>"]
readme = "README.md"
packages =[{include = "pepsflow", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
matplotlib = "^3.9.2"
scienceplots = "^2.1.1"
click = "^8.1.7"
rich = "^13.9.2"
opt-einsum = "^3.4.0"
pytorch-minimize = "^0.0.2"

[tool.poetry.scripts]
pepsflow = "pepsflow.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
